@@ -1,5 +1,5 @@
 {
-  "name": "moviebay",
+  "name": "front",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,26 +8,24 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "core-js": "^3.6.5",
-    "dayjs": "1.10.4",
-    "valine": "^1.4.14",
-    "video.js": "^7.11.8",
-    "vue": "^2.6.12",
-    "vue-router": "^3.5.1",
-    "vuetify": "^2.5.0",
-    "vuex": "^3.6.2"
+    "axios": "^1.2.1",
+    "core-js": "^3.8.3",
+    "dayjs": "^1.11.6",
+    "valine": "^1.5.1",
+    "video.js": "^7.20.3",
+    "vue": "^3.2.45",
+    "vue-router": "^4.1.6",
+    "vuetify": "^3.0.3",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^7.0.0",
-    "vue-template-compiler": "^2.6.12",
-    "vue-cli-plugin-vuetify": "^2.4.0",
-    "vuetify-loader": "^1.7.2"
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3"
   },
   "eslintConfig": {
     "root": true,
@@ -35,16 +33,18 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser"
     },
     "rules": {}
   },
   "browserslist": [
     "> 1%",
-    "last 2 versions"
+    "last 2 versions",
+    "not dead",
+    "not ie 11"
   ]
 }
