@@ -1,14 +1,11 @@
-import Vue from 'vue';
-import Vuetify from 'vuetify/lib';
-
-Vue.use(Vuetify);
+import {createVuetify} from 'vuetify'
 
 /**
  * @see https://vuetifyjs.com/zh-Hans/getting-started/pre-made-layouts
  * @see https://v15.vuetifyjs.com/zh-Hans/framework/grid
  * @see https://vuetifyjs.com/zh-Hans/components/grids
  */
-export default new Vuetify({
+export default createVuetify({
     icons: {
         iconfont: 'md', // fa|md|mdi(default)
     },
