@@ -1,5 +1,12 @@
-module.exports = {
-    transpileDependencies: ['vuetify'],
+const {defineConfig} = require('@vue/cli-service')
+module.exports = defineConfig({
+    transpileDependencies: true,
+    chainWebpack: config => {
+        config.plugin('html').tap(args => {
+            args[0].title = 'MovieBay - 影视湾'
+            return args
+        })
+    },
     devServer: {
         proxy: {
             '/api': {
@@ -11,4 +18,4 @@ module.exports = {
             }
         }
     },
-};
+})
